apps/profile-analyzer/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.20
openai>=1.50.0
playwright>=1.49.0
python-docx>=1.1.0
pdfplumber>=0.11.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
aiofiles>=24.1.0
pillow>=11.0.0
httpx>=0.28.0
beautifulsoup4>=4.12.0
lxml>=5.3.0
